requirements.txt
django>=4.1
psycopg2
gunicorn
djangorestframework
django-cors-headers
django-filter
hgvs
